Admin Dashboard functionallity

+++ backend/controllers/categoryController.js
@@ -1,11 +1,13 @@
 import Category from "../models/categorySchema.js";
+import Product from "../models/productSchema.js";
+import Subcategory from "../models/subcategorySchema.js";
 
 export const getCategory = async (req, res) => {
   try {
     const category = await Category.find();
     return res.status(200).json({
-      success:true,
-      data:category
+      success: true,
+      data: category,
     });
   } catch (err) {
     console.log("Error", err);
@@ -36,19 +38,22 @@ export const createCategory = async (req, res) => {
   try {
     const { name, description } = req.body;
 
-    if(!name || !name.trim())
-    {
-        return res.status(500).json({
-            message:" Category Name is required "
-
-        })
+    if (!name || !name.trim()) {
+      return res.status(500).json({
+        message: " Category Name is required ",
+      });
     }
-    
-    const newCategory = await Category.create({ name:name.trim(), description });
-    return res.status(201).json({success:true,
+
+    const newCategory = await Category.create({
+      name: name.trim(),
+      description,
+    });
+    return res.status(201).json({
+      success: true,
 
       message: "Category Created Successfully ",
-      newCategory
+    
+      data:newCategory
     });
   } catch (err) {
     console.log("Error", err);
@@ -61,24 +66,23 @@ export const updateCategory = async (req, res) => {
     const id = req.params.id;
     const { name, description } = req.body;
 
-  
+    const category_result = await Category.findByIdAndUpdate(
+      id,
+      {
+        name,
+        description,
+      },
+      {
+        new: true,
+        runValidators: true,
+      },
+    );
 
-    const category_result = await Category.findByIdAndUpdate(id, {
-      name,
-      description,
-    },{
-        new:true,
-        runValidators:true
+    if (!category_result) {
+      return res.status(403).json({
+        message: "Category not found",
+      });
     }
-
-);
-
-   if(!category_result)
-   {
-    return res.status(403).json({
-        message:"Category not found"
-    })
-   }
 
     return res.status(200).json({
       success: true,
@@ -93,18 +97,52 @@ export const updateCategory = async (req, res) => {
 export const deleteCategory = async (req, res) => {
   try {
     const id = req.params.id;
+    if (!id) {
+      return res.status(404).json({
+        success: false,
+        message: "Id is required",
+      });
+    }
+
+    await Subcategory.deleteMany({
+      category: id,
+    });
+
     const result = await Category.findByIdAndDelete(id);
-    if(!result)
-   {
-    return res.status(403).json({
-        message:"Category not found"
-    })
-   }
+    if (!result) {
+      return res.status(403).json({
+        message: "Category not found",
+      });
+    }
+
     return res.status(200).json({
       message: "Successfully deleted",
     });
   } catch (err) {
     console.log(err);
     return res.json({ message: err.message });
+  }
+};
+
+export const getAllProductOfCategory = async (req, res) => {
+  try {
+    const category_id = req.params.id;
+    if (!category_id) {
+      return res.status(404).json({
+        success: false,
+        message: "Id is required",
+      });
+    }
+    const product = await Product.find({ category: category_id });
+
+    return res.status(200).json({
+      success: true,
+      data: product,
+    });
+  } catch (err) {
+    return res.status(500).json({
+      success: false,
+      message: err.message,
+    });
   }
 };
